LayoutTests/fast/dom/Window/Location/window-shadow-location-using-string-expected.txt

PASS result is 'pass.html'
